+++ e2e/wallets/integration-wallet-package-keepkey/src/index.ts
@@ -1,6 +1,5 @@
 /*
     E2E testing
-
  */
 
 require("dotenv").config()