+++ apps/pioneer/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @internal/playground
 
+## 0.4.8
+
+### Patch Changes
+
+- bump
+
+## 0.4.7
+
+### Patch Changes
+
+- bump
+
+- bump
+
 ## 0.4.6
 
 ### Patch Changes