+++ e2e/wallets/integration-wallet-package-keepkey/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @coinmasters/integration-wallet-keepkey
 
+## 0.2.8
+
+### Patch Changes
+
+- bump
+
+- Updated dependencies []:
+  - @coinmasters/wallet-keepkey@10.0.9
+  - @coinmasters/types@4.4.8
+  - @coinmasters/core@8.0.8
+
+## 0.2.7
+
+### Patch Changes
+
+- bump
+
+- bump
+
+- Updated dependencies []:
+  - @coinmasters/wallet-keepkey@10.0.8
+  - @coinmasters/types@4.4.7
+  - @coinmasters/core@8.0.7
+
 ## 0.2.6
 
 ### Patch Changes