+++ packages/wallets/metamask/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @coinmasters/wallet-metamask
 
+## 7.0.8
+
+### Patch Changes
+
+- bump
+
+- Updated dependencies []:
+  - @coinmasters/toolbox-cosmos@8.0.8
+  - @coinmasters/toolbox-utxo@8.0.8
+  - @coinmasters/toolbox-evm@8.0.8
+
+## 7.0.7
+
+### Patch Changes
+
+- bump
+
+- bump
+
+- Updated dependencies []:
+  - @coinmasters/toolbox-cosmos@8.0.7
+  - @coinmasters/toolbox-utxo@8.0.7
+  - @coinmasters/toolbox-evm@8.0.7
+
 ## 7.0.6
 
 ### Patch Changes